文字コード変更、ptidをbigintに修正

+++ sql/4.8/orcadbup_teikiprt_4.8.0-1.sql
@@ -1,6 +1,6 @@
 ---------------------
--- ’èŠú¿‹ŠÖŒWƒe[ƒuƒ‹€–Ú‘®«•ÏX
--- ’èŠú¿‹ˆóüî•ñ
+-- Äê´üÀÁµá´Ø·¸¥Æ¡¼¥Ö¥ë¹àÌÜÂ°À­ÊÑ¹¹
+-- Äê´üÀÁµá°õºş¾ğÊó
 ---------------------
 
 \set ON_ERROR_STOP
@@ -9,4 +9,4 @@
 -- tbl_teikiprt --
 alter table tbl_teikiprt alter column hospnum type smallint;
 alter table tbl_teikiprt alter column denpnum type integer;
-alter table tbl_teikiprt alter column ptid    type smallint;
+alter table tbl_teikiprt alter column ptid    type bigint;
